backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.10
django-model-utils==4.3.1
django-widget-tweaks==1.4.12
idna==3.4
judge0api==0.1.20
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
